moved digestFromByteString from Crypto.Hash to Crypto.Hash.Types, Read instance for Digest a
